Dev

See merge request informatik/qe/swess23/group1/g1t2!127

+++ src/main/java/at/qe/g1t2/ui/controllers/UserDetailController.java
@@ -2,9 +2,15 @@ package at.qe.g1t2.ui.controllers;
 
 import at.qe.g1t2.model.Userx;
 import at.qe.g1t2.repositories.LogInfoRepository;
+import at.qe.g1t2.repositories.UserxRepository;
+import at.qe.g1t2.services.SensorStationService;
 import at.qe.g1t2.services.UserService;
+import jakarta.faces.application.FacesMessage;
+import jakarta.faces.context.FacesContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 import java.io.Serializable;
@@ -23,6 +29,11 @@ public class UserDetailController implements Serializable {
     private UserService userService;
     @Autowired
     private LogInfoRepository logInfoRepository;
+    @Autowired
+    private UserxRepository userRepository;
+    @Autowired
+    private SensorStationService sensorStationService;
+    private boolean authenticatedUser;
     /**
      * Attribute to cache the currently displayed user
      */
@@ -69,8 +80,26 @@ public class UserDetailController implements Serializable {
      * Action to delete the currently displayed user.
      */
     public void doDeleteUser() {
-        this.userService.deleteUser(user);
+        if(user==getAuthenticatedUser()){
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_ERROR, "You are not able to delete yourself.", null));
+        }else{
+            this.userService.deleteUser(user);
+        }
         user = null;
+    }
+
+    public boolean isAuthenticatedUser() {
+        return user==getAuthenticatedUser();
+    }
+
+    private Userx getAuthenticatedUser() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        return userRepository.findFirstByUsername(auth.getName());
+    }
+
+    public void setAuthenticatedUser(boolean authenticatedUser) {
+        this.authenticatedUser = authenticatedUser;
     }
 
 }
